+++ src/resources/views/widgets/table_actions.blade.php
@@ -1,13 +1,13 @@
 @isset($actions['edit'])
     @can($actions['edit']['can'])
-        <a href="{{ route($actions['edit']['route'], $model) }}"
+        <a href="{{ route($actions['edit']['route'], $actions['edit']['route_parameters'] ?? $model) }}"
            class="btn btn-xs btn-outline-primary btn-show-on-tr-hover float-right">{{ __('Edit') }}</a>
     @endcan
 @endisset
 
 @isset($actions['delete'])
     @can($actions['delete']['can'])
-        {!! Form::open(['route' => [$actions['delete']['route'], $model],
+        {!! Form::open(['route' => [$actions['delete']['route'], $actions['delete']['route_parameters'] ?? $model],
                         'method' => 'DELETE',
                         'data-confirmation-text' => $actions['delete']['confirmation_text']
                         ])